Removed lettrine's \eqnewif
The package lettrine uses \eqnewif incorrectly as

\eqnewif{\ifL@RTL}{\if@RTL}

Both of \ifL@RTL and \if@RTL are defined: the former one is equivalent to \iffalse by lettrine and the latter one is set to true by RTLdocument option of bidi package but lettrine's use of \eqnewif sets the value of \if@RTL to false and this messes up document typesetting direction.

To fix this, we remove lettrine begindocument hook and then readd them but instead \eqnewif, we simply use

\let\ifL@RTL\if@RTL

which unlike \eqnewif does not cause any issues.

The \eqnewif should be used to equate two new booleans that have not been already defined. Otherwise, if one boolean is set true and the other to false, how are we going to make them equivalent to each other?

This addresses kvafa/bidi#31